Add exception handling for service layer.

+++ src/main/Java/cn/edu/sysu/Service/Impl/OrderFormServiceImpl.java
@@ -1,12 +1,18 @@
 package cn.edu.sysu.Service.Impl;
 
 import cn.edu.sysu.Dao.OrderFormDao;
+import cn.edu.sysu.Dao.RoomDao;
+import cn.edu.sysu.Dto.OperationStatus;
 import cn.edu.sysu.Entity.OrderForm;
+import cn.edu.sysu.Exception.KTVException;
+import cn.edu.sysu.Exception.OrderFormException;
 import cn.edu.sysu.Service.OrderFormService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
@@ -20,61 +26,72 @@ import java.util.List;
 @Service
 public class OrderFormServiceImpl implements OrderFormService {
 
+    //日志对象
+    private Logger logger = LoggerFactory.getLogger(this.getClass());
+
     @Autowired
     private OrderFormDao orderFormDao;
 
+    @Autowired
+    private RoomDao roomDao;
+
     @Override
-    public String addOrderForm(OrderForm OrderForm) {
-        try{
-            orderFormDao.addOrderForm(OrderForm);
-        }catch(Exception e){
-            return "订单新增失败，请联系管理员";
+    @Transactional
+    public OperationStatus addOrderForm(OrderForm orderForm) throws KTVException {
+        try {
+            if (roomDao.queryRoom(orderForm.getRoom().charAt(1) - '0', String.valueOf(orderForm.getRoom().charAt(0))) == null) {
+                throw new OrderFormException("房间不存在，订单无效！");
+            }
+            orderFormDao.addOrderForm(orderForm);
+            return new OperationStatus("新增订单成功！");
+        } catch (Exception e) {
+            logger.error(e.getMessage(), e);
+            throw new KTVException("Inner error:" + e.getMessage());
         }
-        return "订单新增已完成";
     }
 
     @Override
-    public String deleteOrderForm(Date orderTime, String room) {
-        try{
-            orderFormDao.deleteOrderForm(orderTime,room);
-        }catch(Exception e){
-            return "订单删除失败，请联系管理员";
+    @Transactional
+    public OperationStatus deleteOrderForm(Date orderTime, String room) throws KTVException {
+        try {
+            if (roomDao.queryRoom(room.charAt(1) - '0', String.valueOf(room.charAt(0))) == null) {
+                throw new OrderFormException("房间不存在，订单无效！");
+            }
+            orderFormDao.deleteOrderForm(orderTime, room);
+            return new OperationStatus("删除订单成功！");
+        } catch (OrderFormException e1) {
+            throw e1;
+        } catch (Exception e) {
+            logger.error(e.getMessage(), e);
+            throw new KTVException("Inner error:" + e.getMessage());
         }
-        return "订单删除已完成";
     }
 
     @Override
     public List<OrderForm> queryOrderFormByVIP(String phone) {
-        List<OrderForm> result = new ArrayList<>();
-        try{
-            result = orderFormDao.queryOrderFormByVIP(phone);
-        }catch(Exception e){
-            return result;
-        }
-        return result;
-
+        return orderFormDao.queryOrderFormByVIP(phone);
     }
 
     @Override
     public List<OrderForm> queryOrderFormByPay(int pay) {
-        List<OrderForm> result = new ArrayList<>();
-        try{
-            result = orderFormDao.queryOrderFormByPay(pay);
-        }catch(Exception e){
-            return result;
-        }
-        return result;
+        return orderFormDao.queryOrderFormByPay(pay);
     }
 
     @Override
-    public String payOrder(Date orderTime, String room) {
-        try{
-            orderFormDao.payOrder(orderTime,room);
-        }catch (Exception e){
-            return "订单支付失败，请联系管理员";
+    @Transactional
+    public OperationStatus payOrder(Date orderTime, String room) throws KTVException {
+        try {
+            if (roomDao.queryRoom(room.charAt(1) - '0', String.valueOf(room.charAt(0))) == null) {
+                throw new OrderFormException("房间不存在，订单无效！");
+            }
+            orderFormDao.payOrder(orderTime, room);
+            return new OperationStatus("支付订单成功！");
+        } catch (OrderFormException e1) {
+            throw e1;
+        } catch (Exception e) {
+            logger.error(e.getMessage(), e);
+            throw new KTVException("Inner error:" + e.getMessage());
         }
-
-        return "订单支付已完成";
     }
 
 }
